Add comments for Position event, and check for 0 durations
Linear messages hate 0 durations (to the tune of an
OverflowException), so don't send those. This is a bug in the Buttplug
library.

+++ LaunchServerLib/VAMLaunchServer.cs
@@ -5,9 +5,12 @@ namespace VAMLaunch
 {
     public class PositionUpdateEventArgs : EventArgs
     {
+        // 0-99, in Launch FW 1.2 message units
         public uint Position;
+        // 0-99, in Launch FW 1.2 message units
         public uint Speed;
-        public uint Duration;
+        // time in decimal seconds
+        public float Duration;
     }
